requirements.txt
Flask==3.1.1
flask-cors==6.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.1
python-docx==1.2.0
pymupdf==1.26.4
pillow==11.3.0
langchain-google-genai==2.1.10
langchain==0.3.27
langchain-community==0.3.29
pyttsx3==2.99
faiss-cpu==1.12.0